backend/requirements.txt
appdirs==1.4.3
arrow==0.15.4
asgiref==3.2.3
astroid==2.0.4
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.1
black==19.10b0
boltons==19.3.0
bs4==0.0.1
certifi==2018.8.24
cffi==1.13.2
cfgv==3.0.0
chainmap==1.0.3
chardet==3.0.4
Click==7.0
configobj==5.0.6
cryptography==2.8
defusedxml==0.5.0
Django==2.2.8
django-allauth==0.37.1
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-knox==4.1.0
django-smart-selects==1.5.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-base64==2.0
emoji==0.5.4
entrypoints==0.3
flake8==3.7.9
gunicorn==19.7.1
identify==1.4.11
idna==2.7
importlib-metadata==1.5.0
importlib-resources==1.0.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.2.0
nodeenv==1.3.5
oauthlib==2.1.0
pathspec==0.6.0
pbr==5.4.3
pep8==1.7.1
Pillow==5.3.0
pip-autoremove==0.9.1
ppretty==1.3
pre-commit==2.1.0
prettytable==0.7.2
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pyenchant==2.0.0
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.1.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==5.1.2
regex==2019.11.1
requests==2.21.0
requests-oauthlib==1.0.0
ruamel.std.argparse==0.8.1
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
ruamel.yaml.cmd==0.5.5
ruamel.yaml.convert==0.3.2
six==1.11.0
soupsieve==1.9.4
sqlparse==0.3.0
stevedore==1.31.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.24.2
virtualenv==16.7.5
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
whitenoise==3.3.1
wrapt==1.10.11
zipp==3.0.0
